tasks/requirements.txt
PyJWT==1.4.0
amqp==1.4.6
anyjson==0.3.3
billiard==3.3.0.20
celery==3.1.18
cffi==1.1.2
cryptography==0.9.3
enum34==1.0.4
idna==2.0
ipaddress==1.0.14
jira==0.50
kombu==3.0.26
oauthlib==1.0.0
ordereddict==1.1
pyasn1==0.1.8
pycparser==2.14
pycrypto==2.6.1
python-dateutil==2.4.2
pytz==2015.4
requests==2.7.0
requests-oauthlib==0.5.0
requests-toolbelt==0.4.0
six==1.9.0
tlslite==0.4.8
wsgiref==0.1.2
